Write REG_BINARY Values from  a Hex String
When looking for an alternative to AutoIt RegWrite we found this site-package was useful and did all that we wanted except did not appear to work when we wanted to write REG_BINARY values to the registry.  We have been able to modify the existing write_entry() to allow us to write Hex Strings the same way we used to in AutoIt RegWrite using Robotframework keywords.